+++ lib/src/eval/compiler/program.dart
@@ -78,6 +78,9 @@ class Program {
   Uint8List write() {
     final b = BytesBuilder(copy: false);
 
+    b.add([0x45, 0x56, 0x43, 0x00]); // EVC\0
+    b.add(Evc.i32b(64)); // version
+
     _writeMetaBlock(b, topLevelDeclarations.map((key, value) => MapEntry(key.toString(), value)));
     _writeMetaBlock(b, instanceDeclarations.map((key, value) => MapEntry(key.toString(), value)));
     //_writeMetaBlock(b, typeNames);